Navigation Module - version bump:     - kotlin 2.0.0-RC2     - compose 1.6.10-rc01     - decompose 3.0.0     - essenty 2.0.0     - orbitmvi 7.1.0 - minor codestyle fixes - package fixes for login, home modules - added navigation module with StackComponent - changed hardcoded factories to DI in RootComponentImpl
