+++ build-logic/convention/src/main/kotlin/com/archipelago/jobbsyy/MultiplatformFeature.kt
@@ -13,6 +13,7 @@ fun KotlinMultiplatformExtension.configureMultiplatformFeature(
     sourceSets.apply {
         commonMain.dependencies {
             implementation(project(":core:arch"))
+            implementation(project(":core:navigation"))
 
             implementation(libs.napier)
             implementation(libs.kotlinx.coroutines.core)
