+++ build-logic/convention/src/main/kotlin/plugins/ApplicationConventionPlugin.kt
@@ -14,7 +14,8 @@ class ApplicationConventionPlugin : Plugin<Project> {
     override fun apply(target: Project) = with(target) {
         with(pluginManager) {
             apply(libs.plugins.multiplatform.get().pluginId)
-            apply(libs.plugins.compose.get().pluginId)
+            apply(libs.plugins.compose.library.get().pluginId)
+            apply(libs.plugins.compose.compiler.get().pluginId)
             apply(libs.plugins.android.application.get().pluginId)
             apply(libs.plugins.buildConfig.get().pluginId)
             apply(libs.plugins.kotlinx.serialization.get().pluginId)
