+++ build-logic/convention/src/main/kotlin/plugins/ComposeLibraryConventionPlugin.kt
@@ -15,7 +15,8 @@ class ComposeLibraryConventionPlugin : Plugin<Project> {
         with(pluginManager) {
             apply(libs.plugins.multiplatform.get().pluginId)
             apply(libs.plugins.android.library.get().pluginId)
-            apply(libs.plugins.compose.get().pluginId)
+            apply(libs.plugins.compose.library.get().pluginId)
+            apply(libs.plugins.compose.compiler.get().pluginId)
         }
 
         extensions.configure<KotlinMultiplatformExtension> {
