+++ build-logic/convention/src/main/kotlin/plugins/FeatureConventionPlugin.kt
@@ -17,7 +17,8 @@ class FeatureConventionPlugin : Plugin<Project> {
             apply(libs.plugins.multiplatform.get().pluginId)
             apply(libs.plugins.android.library.get().pluginId)
             apply(libs.plugins.buildConfig.get().pluginId)
-            apply(libs.plugins.compose.get().pluginId)
+            apply(libs.plugins.compose.library.get().pluginId)
+            apply(libs.plugins.compose.compiler.get().pluginId)
             apply(libs.plugins.kotlinx.serialization.get().pluginId)
         }
