+++ composeApp/src/commonMain/kotlin/com/archipelago/jobbsyy/App.kt
@@ -1,7 +1,7 @@
 package com.archipelago.jobbsyy
 
 import androidx.compose.runtime.Composable
-import com.archipelago.jobbsyy.root.mvi.RootComponent
+import com.archipelago.jobbsyy.root.component.RootComponent
 import com.archipelago.jobbsyy.root.ui.RootScreen
 import com.archipelago.jobbsyy.theme.AppTheme
