+++ composeApp/src/commonMain/kotlin/com/archipelago/jobbsyy/di/AppModule.kt
@@ -1,7 +1,7 @@
 package com.archipelago.jobbsyy.di
 
-import com.archipelago.jobbsyy.root.di.HomeModule
-import com.archipelago.jobbsyy.root.di.LoginModule
+import com.archipelago.jobbsyy.home.di.HomeModule
+import com.archipelago.jobbsyy.login.di.LoginModule
 import com.archipelago.jobbsyy.root.di.RootModule
 import org.koin.dsl.module
