+++ core/arch/src/commonMain/kotlin/com/archipelago/jobbsyy/arch/StateComponent.kt
@@ -5,5 +5,6 @@ import org.orbitmvi.orbit.ContainerHost
 
 interface StateComponent<STATE : BaseState, SIDE_EFFECT : BaseSideEffect, EVENT : BaseEvent>
     : ComponentContext, ContainerHost<STATE, SIDE_EFFECT> {
+
     fun onEvent(event: EVENT)
 }
