+++ core/navigation/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+    alias(libs.plugins.multiplatform)
+    alias(libs.plugins.kotlinx.serialization)
+}
+
+kotlin {
+    jvm()
+
+    sourceSets {
+        commonMain.dependencies {
+            implementation(libs.decompose)
+            implementation(libs.decompose.compose)
+
+            implementation(project.dependencies.platform(libs.koin.bom))
+            implementation(libs.koin.core)
+        }
+    }
+}
