+++ core/navigation/src/commonMain/kotlin/com/archipelago/jobbsyy/navigation/BaseChild.kt
@@ -1,0 +1,3 @@
+package com.archipelago.jobbsyy.navigation
+
+interface BaseChild
