+++ feature/home/src/commonMain/kotlin/com/archipelago/jobbsyy/home/component/HomeComponent.kt
@@ -1,0 +1,5 @@
+package com.archipelago.jobbsyy.home.component
+
+import com.archipelago.jobbsyy.arch.StateComponent
+
+interface HomeComponent : StateComponent<HomeState, Nothing, Nothing>
