+++ feature/home/src/commonMain/kotlin/com/archipelago/jobbsyy/home/component/HomeState.kt
@@ -1,4 +1,4 @@
-package com.archipelago.jobbsyy.root.mvi
+package com.archipelago.jobbsyy.home.component
 
 import com.archipelago.jobbsyy.arch.BaseState
