--- feature/home/src/commonMain/kotlin/com/archipelago/jobbsyy/root/mvi/HomeComponent.kt
@@ -1,5 +1,0 @@
-package com.archipelago.jobbsyy.root.mvi
-
-import com.archipelago.jobbsyy.arch.StateComponent
-
-interface HomeComponent : StateComponent<HomeState, Nothing, Nothing>
