+++ feature/login/src/commonMain/kotlin/com/archipelago/jobbsyy/login/component/LoginComponent.kt
@@ -1,0 +1,5 @@
+package com.archipelago.jobbsyy.login.component
+
+import com.archipelago.jobbsyy.arch.StateComponent
+
+interface LoginComponent : StateComponent<LoginState, Nothing, LoginEvent>
