+++ feature/login/src/commonMain/kotlin/com/archipelago/jobbsyy/login/component/LoginState.kt
@@ -1,4 +1,4 @@
-package com.archipelago.jobbsyy.root.mvi
+package com.archipelago.jobbsyy.login.component
 
 import com.archipelago.jobbsyy.arch.BaseState
