--- feature/login/src/commonMain/kotlin/com/archipelago/jobbsyy/root/di/LoginModule.kt
@@ -1,9 +1,0 @@
-package com.archipelago.jobbsyy.root.di
-
-import com.archipelago.jobbsyy.root.mvi.LoginComponentImpl
-import org.koin.core.module.dsl.factoryOf
-import org.koin.dsl.module
-
-val LoginModule = module {
-    factoryOf(::LoginComponentImpl)
-}
