--- feature/login/src/commonMain/kotlin/com/archipelago/jobbsyy/root/mvi/LoginComponent.kt
@@ -1,5 +1,0 @@
-package com.archipelago.jobbsyy.root.mvi
-
-import com.archipelago.jobbsyy.arch.StateComponent
-
-interface LoginComponent : StateComponent<LoginState, Nothing, LoginEvent>
