+++ feature/root/src/commonMain/kotlin/com/archipelago/jobbsyy/root/di/RootModule.kt
@@ -1,7 +1,7 @@
 package com.archipelago.jobbsyy.root.di
 
-import com.archipelago.jobbsyy.root.mvi.RootComponent
-import com.archipelago.jobbsyy.root.mvi.RootComponentImpl
+import com.archipelago.jobbsyy.root.component.RootComponent
+import com.archipelago.jobbsyy.root.component.RootComponentImpl
 import org.koin.core.module.dsl.factoryOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
