+++ feature/root/src/commonMain/kotlin/com/archipelago/jobbsyy/root/ui/RootScreen.kt
@@ -1,7 +1,9 @@
 package com.archipelago.jobbsyy.root.ui
 
 import androidx.compose.runtime.Composable
-import com.archipelago.jobbsyy.root.mvi.RootComponent
+import com.archipelago.jobbsyy.home.ui.HomeScreen
+import com.archipelago.jobbsyy.login.ui.LoginScreen
+import com.archipelago.jobbsyy.root.component.RootComponent
 import com.arkivanov.decompose.extensions.compose.stack.Children
 import com.arkivanov.decompose.extensions.compose.stack.animation.slide
 import com.arkivanov.decompose.extensions.compose.stack.animation.stackAnimation
@@ -9,14 +11,12 @@ import com.arkivanov.decompose.extensions.compose.stack.animation.stackAnimation
 @Composable
 fun RootScreen(
     rootComponent: RootComponent
+) = Children(
+    stack = rootComponent.childStack,
+    animation = stackAnimation(slide())
 ) {
-    Children(
-        stack = rootComponent.childStack,
-        animation = stackAnimation(slide())
-    ) {
-        when (val child = it.instance) {
-            is RootComponent.Child.LoginChild -> LoginScreen(child.component)
-            is RootComponent.Child.HomeChild -> HomeScreen(child.component)
-        }
+    when (val child = it.instance) {
+        is RootComponent.Child.LoginChild -> LoginScreen(child.component)
+        is RootComponent.Child.HomeChild -> HomeScreen(child.component)
     }
 }
