+++ settings.gradle.kts
@@ -23,6 +23,7 @@ dependencyResolutionManagement {
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
 include(":core:arch")
+include(":core:navigation")
 include(":core:design-system")
 include(":feature:root")
 include(":feature:login")
